+++ GameDatabase/Entities/Card.cs
@@ -3,8 +3,8 @@
     public partial class Card
     {
         public string AccessCode { get; set; } = null!;
-        public ulong Baid { get; set; }
-        
+        public uint Baid { get; set; }
+
         public virtual UserDatum? Ba { get; set; }
     }
 }
