+++ SharedProject/Models/Requests/SetFavoriteRequest.cs
@@ -2,7 +2,7 @@
 
 public class SetFavoriteRequest
 {
-    public ulong Baid { get; set; }
+    public uint Baid { get; set; }
     public uint SongId { get; set; }
     public bool IsFavorite { get; set; }
 }
