+++ TaikoLocalServer/Services/Interfaces/ISongBestDatumService.cs
@@ -5,13 +5,13 @@ namespace TaikoLocalServer.Services.Interfaces;
 
 public interface ISongBestDatumService
 {
-    public Task<List<SongBestDatum>> GetAllSongBestData(ulong baid);
+    public Task<List<SongBestDatum>> GetAllSongBestData(uint baid);
 
-    public Task<SongBestDatum?> GetSongBestData(ulong baid, uint songId, Difficulty difficulty);
+    public Task<SongBestDatum?> GetSongBestData(uint baid, uint songId, Difficulty difficulty);
 
     public Task UpdateSongBestData(SongBestDatum datum);
 
     public Task InsertSongBestData(SongBestDatum datum);
 
-    public Task<List<SongBestData>> GetAllSongBestAsModel(ulong baid);
+    public Task<List<SongBestData>> GetAllSongBestAsModel(uint baid);
 }
